# tests/unit_tests/fixtures/openapi.yaml
openapi: 3.0.0
info:
  title: V0 Semantic layer API
  version: 1.0.11
  contact:
    name: Андрей Абакин
    url: https://addressbook/search/profile/1763632
    email: abakin.a.v@sberbank.ru
  description: Semantic layer API for EPMPlatform
servers:
  - url: https://ci00163863-epmp-sndbx.apps.dev-terra000004-idm.ocp.delta.sbrf.ru
    description: DEV стенд
  - url: https://ci04763767-semlayer.apps.ift-terra000039-edb.ocp.delta.sbrf.ru
    description: IFT стенд
paths:
...
  /api/v0/tenants/{tenantName}/models/{modelName}/synchronization/dataStorages:
    post:
      tags:
        - v0
      summary: Create data storages In Database From Meta
      operationId: createDataStoragesInDatabaseFromMeta
      responses:
        "200":
          description: Обновленные модели
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/synchronizationResult"
                maxItems: 1000
                uniqueItems: true
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/httpValidationError"
    parameters:
      - $ref: "#/components/parameters/modelName"
      - $ref: "#/components/parameters/tenantName"
      - $ref: "#/components/parameters/checkPossibleToDropMode"
  /api/v0/tenants/{tenantName}/models/{modelName}/synchronization/dataStorages/{dataStorageName}:
    post:
      tags:
        - v0
      summary: Create data storages In Database From Meta
      operationId: createDataStorageInDatabaseFromMeta
      responses:
        "200":
          description: Обновленная таблица
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/synchronizationResult"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/httpValidationError"
    parameters:
      - $ref: "#/components/parameters/modelName"
      - $ref: "#/components/parameters/tenantName"
      - $ref: "#/components/parameters/dataStorageName"
...
  /api/v0/tenants/{tenantName}/models/{modelName}/dataStorages/{dataStorageName}:
    get:
      tags:
        - v0
      summary: Get data storage by model name and data storage name
      operationId: getDataStorageByModelNameAndDataStorageName
      responses:
        "200":
          description: Хранилище `dataStorageName` модели `modelName`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dataStorage"
        "404":
          description: No DSO found in Semantic model for the provided `dataStorageName`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
        "500":
          $ref: "#/components/responses/unexpectedError"
    delete:
      tags:
        - v0
      summary: Delete data storage by model name and data storage name
      operationId: deleteDataStorageByModelNameAndDataStorageName
      responses:
        "200":
          description: Data storage `dataStorageName` в модели `modelName` удалено
        "404":
          description: Data storage `dataStorageName` в модели `modelName` не найдено
        "500":
          $ref: "#/components/responses/unexpectedError"
    patch:
      tags:
        - v0
      summary: Change data storage
      operationId: changeDataStorageByModelNameAndDataStorageName
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dataStorage"
          description: Data storage `dataStorageName` в модели `modelName` изменен
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
          description: Data storage `dataStorageName` в модели `modelName` не найден
        "500":
          $ref: "#/components/responses/unexpectedError"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/dataStorageEdit"
    parameters:
      - $ref: "#/components/parameters/dataStorageName"
      - $ref: "#/components/parameters/modelName"
      - $ref: "#/components/parameters/tenantName"
  /api/v0/tenants/{tenantName}/models/{modelName}/dataStorages/:
    get:
      parameters:
        - example: 1
          name: page
          description: Страница (параметр пагинации)
          schema:
            $ref: "#/components/schemas/paginationPageCount"
          in: query
          required: false
        - example: 20
          name: pageSize
          description: Количество строк (параметр пагинации)
          schema:
            $ref: "#/components/schemas/paginationRowCount"
          in: query
          required: false
        - name: sortDirection
          description: Направление сортировки
          schema:
            $ref: "#/components/schemas/sortDirection"
          in: query
          required: false
        - $ref: "#/components/parameters/dataStorageType"
      tags:
        - v0
      summary: Get all data storages by model name
      operationId: getDataStoragesByModelName
      responses:
        "200":
          description: Описание хранилища `DimensionName` в модели `modelName`
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/dataStorage"
                maxItems: 1000
                uniqueItems: true
        "500":
          $ref: "#/components/responses/unexpectedError"
    post:
      tags:
        - v0
      summary: Create Data Storage by model name
      operationId: createDataStorageByModelName
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dataStorage"
          description: Data Storage создано в модели `modelName`
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
          description: Data Storage уже существует в модели `modelName`
        "500":
          $ref: "#/components/responses/unexpectedError"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/dataStorage"
        description: New dataStorage
    parameters:
      - $ref: "#/components/parameters/modelName"
      - $ref: "#/components/parameters/tenantName"
  /api/v0/tenants/{tenantName}/models/{modelName}/dataStorages/byDbObject:
    post:
      tags:
        - v0
      summary: Get dataStorages by model name and databaseObjects
      operationId: getdataStoragesByModelNameAndDataObject
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/dataStorage"
          description: Хранилище данных модели `modelName` по объекту БД
        "500":
          $ref: "#/components/responses/unexpectedError"
      requestBody:
        description: Объект БД
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/dbObjectForSearch"
    parameters:
      - $ref: "#/components/parameters/modelName"
      - $ref: "#/components/parameters/tenantName"
  /api/v0/tenants/{tenantName}/dataStorages/copyToAnotherModel:
    post:
      tags:
        - v0
      summary: Copy data storages to another model
      operationId: copyDataStoragesToAnotherModel
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/respCopyStorage"
          description: Хранилище данных создано в указанных моделях
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/httpValidationError"
          description: Хранилище данных уже существует в модели `modelName`
      parameters:
        - $ref: "#/components/parameters/tenantName"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/requestCopyStorage"
        description: New model for data storage
...
components:
  responses:
    unexpectedError:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/error"
  parameters:
    modelName:
      name: modelName
      description: Уникальный идентификатор модели
      schema:
        $ref: "#/components/schemas/modelName"
      in: path
      required: true
...
    dataStorageName:
      name: dataStorageName
      description: Уникальный идентификатор data storage
      schema:
        $ref: "#/components/schemas/dataStorageName"
      in: path
      required: true
...
    tenantName:
      name: tenantName
      description: Название тенанта
      schema:
        $ref: "#/components/schemas/tenantName"
      in: path
      required: true
...
    checkPossibleToDropMode:
      name: checkPossibleToDropMode
      description: Строка для проверки возможности удаления модели
      schema:
        $ref: "#/components/schemas/checkPossibleToDropMode"
      in: query
      required: false
...
  schemas:
...
    httpValidationError:
      type: object
      additionalProperties: false
      properties:
        detail:
          type: array
          maxItems: 1000
          uniqueItems: true
          items:
            type: object
            additionalProperties: false
            properties:
              loc:
                type: array
                items:
                  type: string
                  maxLength: 255
                  pattern: '^[^\n\t]+$'
                maxItems: 1000
                uniqueItems: true
              msg:
                $ref: "#/components/schemas/msg"
              type:
                type: string
                maxLength: 255
                pattern: '^[^\n\t]+$'
                description: Тип ошибки
...
    error:
      required:
        - message
      type: object
      additionalProperties: false
      properties:
        message:
          description: A human readable error message
          type: string
          pattern: "^[a-zA-Z0-9_]+$"
          minLength: 1
          maxLength: 255
...
    dataStorage:
      allOf:
        - $ref: "#/components/schemas/versionInfo"
        - $ref: "#/components/schemas/dataStorageEdit"
        - type: object
          properties:
            name:
              $ref: "#/components/schemas/dataStorageName"
            aorType:
              $ref: '#/components/schemas/aorType'
            models:
              type: array
              maxItems: 1000
              minItems: 0
              uniqueItems: true
              description: Массив имён моделей, в которых есть dataStorage, со статусом обработки объекта в ней
              items:
                $ref: "#/components/schemas/modelName"
          additionalProperties: false
          required:
            - name
            - type
            - fields
...
    paginationPageCount:
      example: 1
      description: Страница (параметр пагинации)
      type: integer
      minimum: 1
      maximum: 100
...
    synchronizationResult:
      type: object
      additionalProperties: false
      properties:
        tenant:
          $ref: "#/components/schemas/tenantName"
        modelName:
          $ref: "#/components/schemas/modelName"
        objectName:
          $ref: "#/components/schemas/dataStorageName"
        comment:
          type: string
          description: Комментарий к синхронизации
          maxLength: 255
          minLength: 0
          pattern: '^[^\n\t]+$'
        status:
          type: string
          description: Статус синхронизации
          enum:
            - SUCCESS
            - IGNORED
            - FAILURE
...